Initial version of cltrainsum.icn (new version that summarizes).
Initial commit of program to summarize the output of cltrainparse.icn.  Confusingly, I named the initial program cltrainsum, but then realized it would be cleaner to separate the parsing of CLTRAINEW and the analysis of the data.
This initial version of cltrainsum prints out, for each year, the total distance for each activity, plus the average weight for that year.